+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - ADDED favicon
 - ADDED roster - BMS section (available also from Landing page)
 - ADDED roster - display module popularity (progress bar) for each module
+- ADDED admin - track files sizes